added update modal

+++ database/migrations/2022_02_03_092916_create_expenses_table.php
@@ -19,7 +19,7 @@ class CreateExpensesTable extends Migration
             $table->text('description')->nullable();
             $table->string('category');
             $table->string('subcategory');
-            $table->float('amount');
+            $table->double('amount');
             $table->string('receipt_photo')->nullable();
             $table->timestamps();
         });
